+++ packages/gatsby/src/bootstrap/index.js
@@ -15,13 +15,15 @@ import { getBrowsersList } from "../utils/browserslist"
 import { createSchemaCustomization } from "../utils/create-schema-customization"
 import { startPluginRunner } from "../redux/plugin-runner"
 const { store, emitter } = require(`../redux`)
+import { internalActions } from "../redux/actions"
 const loadPlugins = require(`./load-plugins`)
 const loadThemes = require(`./load-themes`)
 const report = require(`gatsby-cli/lib/reporter`)
 import { getConfigFile } from "./get-config-file"
 const tracer = require(`opentracing`).globalTracer()
-const preferDefault = require(`./prefer-default`)
+import { preferDefault } from "./prefer-default"
 import { removeStaleJobs } from "./remove-stale-jobs"
+const reporter = require(`gatsby-cli/lib/reporter`)
 
 // Show stack trace on unhandled promises.
 process.on(`unhandledRejection`, (reason, p) => {
@@ -31,12 +33,12 @@ process.on(`unhandledRejection`, (reason, p) => {
 import { createGraphQLRunner } from "./create-graphql-runner"
 const { extractQueries } = require(`../query/query-watcher`)
 const requiresWriter = require(`./requires-writer`)
-const { writeRedirects } = require(`./redirects-writer`)
+import { writeRedirects, startRedirectListener } from "./redirects-writer"
 
 // Override console.log to add the source file + line number.
 // Useful for debugging if you lose a console.log somewhere.
 // Otherwise leave commented out.
-// require(`./log-line-function`)
+// import "./log-line-function"
 
 type BootstrapArgs = {
   directory: string,
@@ -68,6 +70,8 @@ module.exports = async (args: BootstrapArgs) => {
   // Start plugin runner which listens to the store
   // and invokes Gatsby API based on actions.
   startPluginRunner()
+
+  startRedirectListener()
 
   const directory = slash(args.directory)
 
@@ -154,10 +158,7 @@ module.exports = async (args: BootstrapArgs) => {
     )
   }
 
-  store.dispatch({
-    type: `SET_SITE_CONFIG`,
-    payload: config,
-  })
+  store.dispatch(internalActions.setSiteConfig(config))
 
   activity.end()
 
@@ -444,7 +445,14 @@ module.exports = async (args: BootstrapArgs) => {
     parentSpan: bootstrapSpan,
   })
   activity.start()
-  await require(`../utils/source-nodes`)({ parentSpan: activity.span })
+  await require(`../utils/source-nodes`).default({ parentSpan: activity.span })
+  reporter.verbose(
+    `Now have ${store.getState().nodes.size} nodes with ${
+      store.getState().nodesByType.size
+    } types: [${[...store.getState().nodesByType.entries()]
+      .map(([type, nodes]) => type + `:` + nodes.size)
+      .join(`, `)}]`
+  )
   activity.end()
 
   // Create Schema.
@@ -485,6 +493,13 @@ module.exports = async (args: BootstrapArgs) => {
       parentSpan: activity.span,
     },
     { activity }
+  )
+  reporter.verbose(
+    `Now have ${store.getState().nodes.size} nodes with ${
+      store.getState().nodesByType.size
+    } types, and ${
+      store.getState().nodesByType?.get(`SitePage`).size
+    } SitePage nodes`
   )
   activity.end()