+++ packages/gatsby-theme-blog-core/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.4.15...gatsby-theme-blog-core@1.4.16) (2020-05-06)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.4.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.4.14...gatsby-theme-blog-core@1.4.15) (2020-05-06)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.4.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.4.13...gatsby-theme-blog-core@1.4.14) (2020-05-06)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.4.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.4.12...gatsby-theme-blog-core@1.4.13) (2020-05-05)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.4.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.4.11...gatsby-theme-blog-core@1.4.12) (2020-05-05)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.4.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.4.10...gatsby-theme-blog-core@1.4.11) (2020-05-04)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.4.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.4.9...gatsby-theme-blog-core@1.4.10) (2020-05-04)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
 ## [1.4.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.4.8...gatsby-theme-blog-core@1.4.9) (2020-05-01)
 
 **Note:** Version bump only for package gatsby-theme-blog-core